Add support for no_std environments
This commit adds a default "std" feature that can be disabled to run
this crate in no_std environments. This allows for Zlib compression or
decompression to happen on embedded systems.

Rel: https://github.com/dfrg/swash/issues/4

This is a breaking change, as anyone who is already importing yazi with
"default-features = false" and using the From<io::Error> impl on
yazi::Error will have their build break after this commit.

Signed-off-by: John Nunley <dev@notgull.net>